Move MojangAssetIndex out of the api package
