+++ src/main/java/com/atlauncher/data/mojang/MojangAssetIndex.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.data.mojang.api;
+package com.atlauncher.data.mojang;
 
 import com.atlauncher.annot.Json;
